+++ src/utils/db.ts
@@ -1,5 +1,6 @@
-import { Database } from "bun:sqlite";
 import { resolve } from "path";
+
+import { Database } from "bun:sqlite";
 
 export const database = new Database(resolve(__dirname, "db.db"), {
   create: true,
